+++ src/frame_info.rs
@@ -1,14 +1,4 @@
-use crate::{Frame, MAX_INPUT_BYTES, MAX_PLAYERS, NULL_FRAME};
-
-/// The input buffer used to save the bytes from a player input. It is bigger than necessary by a factor `MAX_PLAYERS` to allow the same type of buffers to be used to transmit
-/// player inputs for all players to the spectators. This definitely isn't optimal and might be changed later.
-pub type InputBuffer = [u8; MAX_INPUT_BYTES * MAX_PLAYERS as usize];
-
-pub const BLANK_INPUT: GameInput = GameInput {
-    frame: NULL_FRAME,
-    buffer: [0; MAX_INPUT_BYTES * MAX_PLAYERS as usize],
-    size: 0,
-};
+use crate::{Frame, NULL_FRAME};
 
 /// Computes the fletcher16 checksum, copied from wikipedia: <https://en.wikipedia.org/wiki/Fletcher%27s_checksum>
 fn fletcher16(data: &[u8]) -> u16 {
@@ -65,14 +55,14 @@ impl GameState {
 
 /// Represents a serialized input for a single player in a single frame. This struct holds a `buffer` where the first `size` bytes represent the encoded input of a single player.
 /// The associated frame is denoted with `frame`. You do not need to create this struct, but the sessions will provide a `Vec<GameInput>` for you during `advance_frame()`.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct GameInput {
     /// The frame to which this info belongs to. -1/`NULL_FRAME` represents an invalid frame
     pub frame: Frame,
     // The input size
     pub size: usize,
     /// An input buffer that will hold input data
-    pub buffer: InputBuffer,
+    pub buffer: Vec<u8>,
 }
 
 impl Default for GameInput {
@@ -86,35 +76,34 @@ impl Default for GameInput {
 }
 
 impl GameInput {
-    pub(crate) fn new(frame: Frame, size: usize) -> Self {
-        assert!(size > 0);
+    pub(crate) fn new(frame: Frame, size: usize, buffer: Vec<u8>) -> Self {
+        assert!(size == buffer.len());
+
         Self {
             frame,
             size,
-            buffer: Default::default(),
+            buffer,
+        }
+    }
+
+    pub(crate) fn blank_input(size: usize) -> Self {
+        Self {
+            frame: NULL_FRAME,
+            size,
+            buffer: vec![0; size],
         }
     }
 }
 
 impl GameInput {
-    pub(crate) fn copy_input(&mut self, bytes: &[u8]) {
-        assert!(bytes.len() == self.size);
-        self.buffer[0..self.size].copy_from_slice(bytes);
-    }
-
     pub(crate) fn erase_bits(&mut self) {
-        self.buffer.iter_mut().for_each(|m| *m = 0)
+        self.buffer = vec![0; self.size];
     }
 
     pub(crate) fn equal(&self, other: &Self, bitsonly: bool) -> bool {
         (bitsonly || self.frame == other.frame)
             && self.size == other.size
             && self.buffer == other.buffer
-    }
-
-    /// Retrieve your serialized input with this method. Returns a slice which you can use to deserialize.
-    pub fn input(&self) -> &[u8] {
-        &self.buffer[0..self.size]
     }
 }
 
@@ -132,10 +121,9 @@ mod game_input_tests {
         let fake_inputs: u32 = 5;
         let input_size = std::mem::size_of::<u32>();
         let serialized_inputs = bincode::serialize(&fake_inputs).unwrap();
-        let mut input1 = GameInput::new(0, input_size);
-        input1.copy_input(&serialized_inputs);
-        let mut input2 = GameInput::new(5, input_size);
-        input2.copy_input(&serialized_inputs);
+        let input1 = GameInput::new(0, input_size, serialized_inputs);
+        let serialized_inputs = bincode::serialize(&fake_inputs).unwrap();
+        let input2 = GameInput::new(5, input_size, serialized_inputs);
         assert!(input1.equal(&input2, true)); // different frames, but does not matter
     }
 
@@ -145,13 +133,11 @@ mod game_input_tests {
 
         let fake_inputs: u32 = 5;
         let serialized_inputs = bincode::serialize(&fake_inputs).unwrap();
-        let mut input1 = GameInput::new(0, input_size);
-        input1.copy_input(&serialized_inputs);
+        let input1 = GameInput::new(0, input_size, serialized_inputs);
 
         let fake_inputs: u32 = 7;
         let serialized_inputs = bincode::serialize(&fake_inputs).unwrap();
-        let mut input2 = GameInput::new(0, input_size);
-        input2.copy_input(&serialized_inputs);
+        let input2 = GameInput::new(0, input_size, serialized_inputs);
 
         assert!(!input1.equal(&input2, false)); // different bits
     }